requirements.txt
pyspark==3.5.0
pandas==2.2.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
mlflow==2.11.3
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pyyaml==6.0.1
pytest==8.0.2
httpx==0.27.0
evidently==0.4.17
numpy==1.26.4
pyarrow==15.0.0
ray[default]>=2.31.0
psutil==5.9.8
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0